[pax] Updated to patch 3.2.39-test97
Updated to pax-linux-3.2.39-test97.patch

PaX Team says:
  - fixed kprobes problem under KERNEXEC/i386, reported by Emese Revfy
  - fixed arm compilation problems reported by Michael Tremer
  - the constify plugin got smarter that enabled, with some additional patching,
    the elimination of about half the static function pointers on amd64/allmod
    (up from about 18%), depending on the kernel config it can be even more (70%)
  - fixed too strict mmap parameter checking on i386, reported by browndav (http://forums.grsecurity.net/viewtopic.php?f=1&t=3339)
  - small latent entropy improvement: pass pax_extra_latent_entropy to the kernel to extract entropy from RAM content during boot
  - adjusted RLIMIT_AS accounting for the extra ASLR gap mappings, reported by Alexander Stoll (https://bugs.gentoo.org/show_bug.cgi?id=459268)
  - changed debian packager rules to include the compiler plugins, from Tyler Coumbes <coumbes@gmail.com>
  - fixed the sa_restorer leak discovered and reported by Emese Revfy (CVE-2013-0914, google chromium bug #177956)
  - new size overflow plugin from Emese that instruments a whole lot more code due to tracking function return values
    and more type casts as well. this found the above mentioned sa_restorer leak and would have protected against CVE-2013-0913.
  - added a small sanity check to kfree (spender's idea)
  - fixed a size overflow false positive (well, not really, it's bad zlib code) reported by slashbeast on irc
  - fixed ptrace compat bug found by the size overflow plugin
  - fixed perf compilation reported by Michael Tremer
  - fixed USERCOPY reports triggered by SCTP, reported by Marc Schiffbauer (https://bugzilla.redhat.com/show_bug.cgi?id=919315#c11)
  - fixed bug with the old PAGEEXEC method and hugetlb, reported by Alex Efros (https://bugs.gentoo.org/show_bug.cgi?id=437722)
  - added a new gcc plugin to plug (pun intended) some of the kernel stack leaks to userland

+++ fs/proc/base.c
@@ -910,7 +910,7 @@ static ssize_t environ_read(struct file *file, char __user *buf,
 
 	ret = 0;
 	while (count > 0) {
-		int this_len, retval, max_len;
+		long this_len, retval, max_len;
 
 		this_len = mm->env_end - (mm->env_start + src);
